CurvePath: Remove factory methods
